refactor(File): Extract ShareContent from Status
