+++ src/drive/web/modules/filelist/cells/Status.jsx
@@ -1,40 +1,15 @@
 import React, { useState } from 'react'
 import cx from 'classnames'
 
-import { SharedStatus, ShareModal } from 'cozy-sharing'
+import { ShareModal } from 'cozy-sharing'
 import { TableCell } from 'cozy-ui/transpiled/react/Table'
 import Icon from 'cozy-ui/transpiled/react/Icon'
 import palette from 'cozy-ui/transpiled/react/palette'
 import PhoneDownloadIcon from 'cozy-ui/transpiled/react/Icons/PhoneDownload'
 
-import HammerComponent from '../HammerComponent'
+import { ShareContent } from 'drive/web/modules/filelist/cells/ShareContent'
 
 import styles from 'drive/styles/filelist.styl'
-
-const ShareContent = ({
-  file,
-  setDisplayedModal,
-  disabled,
-  isInSyncFromSharing
-}) => (
-  <div
-    className={cx(styles['fil-content-sharestatus'], {
-      [styles['fil-content-sharestatus--disabled']]: disabled
-    })}
-  >
-    {isInSyncFromSharing ? (
-      <span data-testid="fil-content-sharestatus--noAvatar">—</span>
-    ) : (
-      <HammerComponent
-        onClick={() => {
-          !disabled && setDisplayedModal(true) // should be only disabled
-        }}
-      >
-        <SharedStatus docId={file.id} />
-      </HammerComponent>
-    )}
-  </div>
-)
 
 const Status = ({
   isAvailableOffline,
